group globs and mounts together
In the previous approach, globs would not be considered if a mounted app
matched the route. On first glance this would be fair, but if that
mounted app is at the root, and the glob is not, then even when
the glob is listed first, it would be ignored (see #262).

But, a glob and a mount are not so different - they both have the
potential to be far-reaching catch-all endpoints (though a glob is
http_method-specific). So let’s handle them in the same collection,
which helps keep their listed order in a routes file. This allows
more-specific globs to be considered before a less-specific mount,
provided they’re ordered that way in the routes.